% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_candidate_new.R
\name{cf_candidate_new}
\alias{cf_candidate_new}
\title{Campaign finance - get new candidates}
\usage{
cf_candidate_new(campaign_cycle = NULL, key = NULL, ...)
}
\arguments{
\item{campaign_cycle}{An even-numbdered year in YYYY format. Presidential data:
2008-present; Congressional data: 2000-present}

\item{key}{your New York Times API key; pass in, or loads from .Renviron as 
\code{NYTIMES_CF_KEY}. See \code{\link{rtimes-package}} for info on rate limits.}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}

\item{state}{(character) Two-letter state abbreviation}

\item{chamber}{(character) house or senate (optional)}

\item{district}{(character) Specify the district number. Use 1 for states 
with a single representative. (House requests only - districts with 
Senate requests will be ignored.)}
}
\description{
Campaign finance - get new candidates
}
\examples{
\dontrun{
cf_candidate_new(campaign_cycle = 2016)
cf_candidate_new(campaign_cycle = 2015)
cf_candidate_new(campaign_cycle = 2014)
}
}
\references{
\url{http://propublica.github.io/campaign-finance-api-docs}
}

