% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cf_candidate_search}
\alias{cf_candidate_search}
\title{Campaign finance NYTimes API - candidate search}
\usage{
cf_candidate_search(campaign_cycle = NULL, query = NULL,
  key = getOption("nytimes_cf_key"), callopts = list())
}
\arguments{
\item{campaign_cycle}{An even-numbdered year in YYYY format. Presidential data: 2008-present; Congressional data: 2000-present}

\item{query}{Last name of a candidate}

\item{key}{your New York Times API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\description{
See \url{http://developer.nytimes.com/docs/campaign_finance_api/}
}
\examples{
\dontrun{
cf_candidate_search(campaign_cycle = 2010, query='smith')
cf_candidate_search(campaign_cycle = 2008, query='obama')
}
}

