% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_candidate_state.R
\name{cf_candidate_state}
\alias{cf_candidate_state}
\title{Campaign finance - candidates from a state}
\usage{
cf_candidate_state(campaign_cycle = NULL, state = NULL, chamber = NULL,
  district = NULL, key = NULL, ...)
}
\arguments{
\item{campaign_cycle}{An even-numbdered year in YYYY format. Presidential data:
2008-present; Congressional data: 2000-present}

\item{state}{(character) Two-letter state abbreviation}

\item{chamber}{(character) house or senate (optional)}

\item{district}{(character) Specify the district number. Use 1 for states 
with a single representative. (House requests only - districts with 
Senate requests will be ignored.)}

\item{key}{your New York Times API key; pass in, or loads from .Renviron as 
\code{NYTIMES_CF_KEY}. See \code{\link{rtimes-package}} for info on rate limits.}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}
}
\description{
Campaign finance - candidates from a state
}
\examples{
\dontrun{
cf_candidate_state(campaign_cycle = 2016, state = "DE")
cf_candidate_state(campaign_cycle = 2016, state = "CA")
}
}
\references{
\url{http://propublica.github.io/campaign-finance-api-docs}
}

