% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cg_membersleaving.R
\name{cg_membersleaving}
\alias{cg_membersleaving}
\title{Get a list of members who have left the Senate or House or have announced plans to do so.}
\usage{
cg_membersleaving(congress_no = NULL, chamber = NULL,
  key = getOption("nytimes_cg_key",
  stop("need an API key for the NYT Congress API")), callopts = list())
}
\arguments{
\item{congress_no}{The number of the Congress during which the members served.}

\item{chamber}{One of 'house' or 'senate.}

\item{key}{your New York Times API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\value{
List of new members of he current Congress.
}
\description{
Get a list of members who have left the Senate or House or have announced plans to do so.
}
\examples{
\dontrun{
cg_membersleaving(112, 'house')
}
}

