% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cg_membersponsorcompare}
\alias{cg_membersponsorcompare}
\title{Compare bill sponsorship between two members who served in the same Congress
and chamber.}
\usage{
cg_membersponsorcompare(memberid_1 = NULL, memberid_2 = NULL,
  congress_no = NULL, chamber = NULL, key = getOption("nytimes_cg_key",
  stop("need an API key for the NYT Congress API")), callopts = list())
}
\arguments{
\item{memberid_1}{The member's unique ID number (alphanumeric). To find a
member's ID number, get the list of members for the appropriate House
or Senate. You can also use the Biographical Directory of the United
States Congress to get a member's ID. In search results, each member's
name is linked to a record by index ID (e.g.,
http://bioguide.congress.gov/scripts/biodisplay.pl?index=C001041).
Use the index ID as member-id in your request.}

\item{memberid_2}{See description for memberid_1.}

\item{congress_no}{The number of the Congress during which the members served.}

\item{chamber}{One of 'house' or 'senate.}

\item{key}{your New York Times API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\value{
Compare bill sponsorship between two members who served in the same
   Congress and chamber.
}
\description{
Compare bill sponsorship between two members who served in the same Congress
and chamber.
}
\examples{
\dontrun{
cg_membersponsorcompare('S001181', 'A000368', 112, 'senate')
}
}

