% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cg_membervotepositions}
\alias{cg_membervotepositions}
\title{Get a list of members who have left the Senate or House or have announced
plans to do so.}
\usage{
cg_membervotepositions(memberid = NULL, key = getOption("nytimes_cg_key",
  stop("need an API key for the NYT Congress API")), callopts = list())
}
\arguments{
\item{memberid}{The member's unique ID number (alphanumeric). To find a
member's ID number, get the list of members for the appropriate House
or Senate. You can also use the Biographical Directory of the United
States Congress to get a member's ID. In search results, each member's
name is linked to a record by index ID (e.g.,
http://bioguide.congress.gov/scripts/biodisplay.pl?index=C001041).
Use the index ID as member-id in your request.}

\item{key}{your New York Times API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\value{
List of new members of he current Congress.
}
\description{
Get a list of members who have left the Senate or House or have announced
plans to do so.
}
\examples{
\dontrun{
cg_membervotepositions('S001181')
}
}

