% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cg_newmembers.R
\name{cg_newmembers}
\alias{cg_newmembers}
\title{Get a list of the most recent new members of the current Congress.}
\usage{
cg_newmembers(key = getOption("nytimes_cg_key",
  stop("need an API key for the NYT Congress API")), ...)
}
\arguments{
\item{key}{your New York Times API key; loads from .Rprofile}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}
}
\value{
List of new members of he current Congress.
}
\description{
Get a list of the most recent new members of the current Congress.
}
\examples{
\dontrun{
cg_newmembers()
}
}

