% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cf_candidate_details.R
\name{cf_candidate_details}
\alias{cf_candidate_details}
\title{Campaign finance NYTimes API - candidate details}
\usage{
cf_candidate_details(campaign_cycle = NULL, fec_id = NULL,
  key = getOption("nytimes_cf_key"), callopts = list())
}
\arguments{
\item{campaign_cycle}{An even-numbdered year in YYYY format. Presidential data: 2008-present; Congressional data: 2000-present}

\item{fec_id}{FEC id}

\item{key}{your New York Times API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\description{
See \url{http://developer.nytimes.com/docs/campaign_finance_api/}
}
\examples{
\dontrun{
cf_candidate_details(campaign_cycle = 2008, fec_id = 'P80003338')
}
}

