% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_candidate_details.R
\name{cf_candidate_details}
\alias{cf_candidate_details}
\title{Campaign finance - candidate details}
\usage{
cf_candidate_details(campaign_cycle = NULL, fec_id = NULL, key = NULL,
  ...)
}
\arguments{
\item{campaign_cycle}{An even-numbdered year in YYYY format. Presidential data:
2008-present; Congressional data: 2000-present}

\item{fec_id}{FEC id}

\item{key}{(character) your Propublica API key; pass in, or loads from .Renviron as 
\code{PROPUBLICA_API_KEY}. See \code{\link{rtimes-package}} for info on rate limits.}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}
}
\description{
Campaign finance - candidate details
}
\examples{
\dontrun{
cf_candidate_details(campaign_cycle = 2008, fec_id = 'P80003338')
cf_candidate_details(campaign_cycle = 2012, fec_id = 'P80003338')
}
}
\references{
\url{http://propublica.github.io/campaign-finance-api-docs}
}
