% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cf_candidate_details.R
\name{cf_candidate_details}
\alias{cf_candidate_details}
\title{Campaign finance NYTimes API - candidate details}
\usage{
cf_candidate_details(campaign_cycle = NULL, fec_id = NULL, key = NULL,
  ...)
}
\arguments{
\item{campaign_cycle}{An even-numbdered year in YYYY format. Presidential data:
2008-present; Congressional data: 2000-present}

\item{fec_id}{FEC id}

\item{key}{your New York Times API key; pass in, or loads from .Rprofile as
\code{nytimes_cf_key}, or from .Renviron as \code{NYTIMES_CF_KEY}}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}
}
\description{
Campaign finance NYTimes API - candidate details
}
\examples{
\dontrun{
cf_candidate_details(campaign_cycle = 2008, fec_id = 'P80003338')
cf_candidate_details(campaign_cycle = 2012, fec_id = 'P80003338')
}
}
\references{
\url{http://developer.nytimes.com/docs/campaign_finance_api/}
}

