% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cf_candidate_search.R
\name{cf_candidate_search}
\alias{cf_candidate_search}
\title{Campaign finance NYTimes API - candidate search}
\usage{
cf_candidate_search(campaign_cycle = NULL, query = NULL, key = NULL, ...)
}
\arguments{
\item{campaign_cycle}{An even-numbdered year in YYYY format. Presidential data:
2008-present; Congressional data: 2000-present}

\item{query}{Last name of a candidate}

\item{key}{your New York Times API key; pass in, or loads from .Rprofile as
\code{nytimes_cf_key}, or from .Renviron as \code{NYTIMES_CF_KEY}}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}
}
\description{
Campaign finance NYTimes API - candidate search
}
\examples{
\dontrun{
# cf_candidate_search(campaign_cycle = 2010, query='smith')
# cf_candidate_search(campaign_cycle = 2008, query='obama')
}
}
\references{
\url{http://developer.nytimes.com/docs/campaign_finance_api/}
}
\keyword{internal}

