% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cg_memberbioroles.R
\name{cg_memberbioroles}
\alias{cg_memberbioroles}
\title{Member bio roles}
\usage{
cg_memberbioroles(memberid = NULL, key = NULL, ...)
}
\arguments{
\item{memberid}{The member's unique ID number (alphanumeric). To find a
member's ID number, get the list of members for the appropriate House
or Senate. You can also use the Biographical Directory of the United
States Congress to get a member's ID. In search results, each member's
name is linked to a record by index ID (e.g.,
http://bioguide.congress.gov/scripts/biodisplay.pl?index=C001041).
Use the index ID as member-id in your request.}

\item{key}{your New York Times API key; pass in, or loads from .Rprofile as
\code{nytimes_cg_key}, or from .Renviron as \code{NYTIMES_CG_KEY}}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}
}
\value{
List of members of a particular chamber in a particular Congress.
}
\description{
Get biographical and Congressional role information for a particular member of Congress.
}
\examples{
\dontrun{
cg_memberbioroles('S001181')
}
}

