% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cg_newmembers}
\alias{cg_newmembers}
\title{Get a list of the most recent new members of the current Congress.}
\usage{
cg_newmembers(key = getOption("nytimes_cg_key",
  stop("need an API key for the NYT Congress API")), callopts = list())
}
\arguments{
\item{key}{your New York Times API key; loads from .Rprofile}

\item{callopts}{Optional additional curl options (debugging tools mostly)}
}
\value{
List of new members of he current Congress.
}
\description{
Get a list of the most recent new members of the current Congress.
}
\examples{
\dontrun{
cg_newmembers()
}
}

