% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_newmembers.R
\name{cg_newmembers}
\alias{cg_newmembers}
\title{Get a list of the most recent new members of the current Congress.}
\usage{
cg_newmembers(key = NULL, ...)
}
\arguments{
\item{key}{(character) your Propublica API key; pass in, or loads from .Renviron
as \code{PROPUBLICA_API_KEY}. See \code{\link{rtimes-package}} for info on rate limits.}

\item{...}{Curl options (debugging tools mostly) passed to \code{\link[httr]{GET}}}
}
\value{
List of new members of he current Congress.
}
\description{
Get a list of the most recent new members of the current Congress.
}
\examples{
\dontrun{
cg_newmembers()
}
}
